+++ packages/core/src/storybook/stand-alone-documentaion/spacing/spacing-token-name/spacing-token-name.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import styles from "./spacing-usage-examples.module.scss";
+
+export const TokenName = ({ children }: { children: React.ReactNode }) => (
+  <span className={styles.tokenName}>{children}</span>
+);